Revert "Revert "test: add condition for kbs deploy in provisioner to align with e2e test""
This reverts commit 6c87f313a9ba980cd4a778a806d4c5c67ea04ecd.

+++ src/cloud-api-adaptor/test/tools/provisioner-cli/main.go
@@ -173,14 +173,16 @@ func main() {
 			log.Fatal(err)
 		}
 
-		props := provisioner.GetProperties(context.TODO(), cfg)
-		keyBrokerService, err := pv.NewKeyBrokerService(props["CLUSTER_NAME"])
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		if err = keyBrokerService.Delete(context.TODO(), cfg); err != nil {
-			log.Fatal(err)
+		if strings.EqualFold(shouldDeployKbs, "true") || strings.EqualFold(shouldDeployKbs, "yes") {
+			props := provisioner.GetProperties(context.TODO(), cfg)
+			keyBrokerService, err := pv.NewKeyBrokerService(props["CLUSTER_NAME"])
+			if err != nil {
+				log.Fatal(err)
+			}
+	
+			if err = keyBrokerService.Delete(context.TODO(), cfg); err != nil {
+				log.Fatal(err)
+			}
 		}
 
 		kubeconfigPath := kconf.ResolveKubeConfigFile()
